requirements.txt
mysql-connector-python==8.0.33
tabulate==0.9.0
user-agents>=2.2.0
